Fix PI memset signatures in HIP and CUDA plugins

+++ sycl/plugins/cuda/pi_cuda.cpp
@@ -5280,7 +5280,7 @@ pi_result cuda_piextUSMEnqueueFill2D(pi_queue, void *, size_t, size_t,
 
 // TODO: Implement this. Remember to return true for
 //       PI_EXT_ONEAPI_CONTEXT_INFO_USM_MEMSET2D_SUPPORT when it is implemented.
-pi_result cuda_piextUSMEnqueueMemset2D(pi_queue, void *, size_t, size_t, int,
+pi_result cuda_piextUSMEnqueueMemset2D(pi_queue, void *, size_t, int, size_t,
                                        size_t, pi_uint32, const pi_event *,
                                        pi_event *) {
   sycl::detail::pi::die("cuda_piextUSMEnqueueMemset2D: not implemented");
